% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{new_treenode}
\alias{new_treenode}
\title{new_treenode}
\usage{
new_treenode(
  parent = integer(),
  id = integer(),
  xmin = double(),
  xmax = double(),
  ymin = double(),
  ymax = double()
)
}
\arguments{
\item{parent}{an integer vector}

\item{id}{an integer vector}

\item{xmin}{a double vector}

\item{xmax}{a double vector}

\item{ymin}{a double vector}

\item{ymax}{a double vector}
}
\description{
new_treenode
}
