% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadnet-package.R
\docType{package}
\name{quadnet-package}
\alias{quadnet}
\alias{quadnet-package}
\title{quadnet: Build, analyse and visualise QuadTree road network models}
\description{
Generate random road networks using the QuadTree model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ppintosilva/quadnet}
  \item Report bugs at \url{https://github.com/ppintosilva/quadnet/issues}
}

}
\keyword{internal}
