% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{treenode}
\alias{treenode}
\title{\code{treenode} vector}
\usage{
treenode(parent, id, xmin, xmax, ymin, ymax)
}
\arguments{
\item{parent}{an integer vector}

\item{id}{an integer vector}

\item{xmin}{a double vector}

\item{xmax}{a double vector}

\item{ymin}{a double vector}

\item{ymax}{a double vector}
}
\value{
As S3 vector of class \code{vctrs_treenode}
}
\description{
This creates a vector of treenodes, each encoding a square in the Cartesian
space.
}
\examples{
# An example tree with:
# - 1 root (box = (0,100,0,100)),
# - 2 childs (box1 = (0,50,0,50), box2 = (50,100,50,100))

treenode(
  parent = c(0,0,0),
  id = 0:2,
  xmin = c(0,0,50),
  xmax = c(100,50,100),
  ymin = c(0,0,50),
  ymax = c(100,50,100)
)
}
